+++ src/Contents/rollup.config.js
@@ -4,8 +4,27 @@ import commonjs from '@rollup/plugin-commonjs';
 import terser from '@rollup/plugin-terser';
 import { defineConfig } from 'rollup';
 import manifestImport from 'rollup-plugin-output-manifest';
+import postcss from 'rollup-plugin-postcss';
+import { spawn } from 'node:child_process';
 
 const outputManifest = manifestImport?.outputManifest ?? manifestImport?.default ?? manifestImport;
+
+function patchManifestPlugin() {
+    return {
+        name: 'ci-patch-manifest',
+        async writeBundle() {
+            return new Promise((resolve, reject) => {
+                const child = spawn(
+                    'node',
+                    ['writable/resources/frontend/scripts/patch-manifest.mjs'],
+                    { stdio: 'inherit' }
+                );
+
+                child.on('exit', (code) => (code === 0 ? resolve() : reject(new Error(`patch-manifest exit code ${code}`))))
+            });
+        }
+    }
+}
 
 export default defineConfig({
     input: {
@@ -25,6 +44,10 @@ export default defineConfig({
             extensions: [".mjs", ".js", ".json", ".ts"]
         }),
         commonjs(),
+        postcss({
+            extract: true,
+            minimize: true,
+        }),
         typescript({ tsconfig: "./tsconfig.json" }),
         terser({
             format: { comments: false },
@@ -34,6 +57,7 @@ export default defineConfig({
         outputManifest({
             fileName: "manifest.json",
             publicPath: "/assets/"
-        })
+        }),
+        patchManifestPlugin(),
     ]
 });